+++ src/test/scala/uk/co/gresearch/spark/dgraph/connector/sources/TestTriplesSource.scala
@@ -103,7 +103,7 @@ class TestTriplesSource extends FunSpec with SparkTestSession {
           .as[StringTriple]
           .collectAsList()
       rows.forEach(println)
-      assert(rows.size() == 34)
+      assert(rows.size() == 44)
     }
 
     it("should encode TypedTriple") {
@@ -115,7 +115,7 @@ class TestTriplesSource extends FunSpec with SparkTestSession {
           .as[TypedTriple]
           .collectAsList()
       rows.forEach(println)
-      assert(rows.size() == 34)
+      assert(rows.size() == 44)
     }
 
     it("should fail without target") {